basic spell fix.
